Car: consolidate fingerprint helpers extensions (#263)
* Car: Fingerprinting v1 extensions

* move to a single helper
