+++ opendbc/car/chrysler/fingerprints.py
@@ -2,7 +2,7 @@
 from opendbc.car.structs import CarParams
 from opendbc.car.chrysler.values import CAR
 
-from opendbc.sunnypilot.car.fw_versions_ext import merge_fw_versions
+from opendbc.sunnypilot.car.fingerprints_ext import merge_fw_versions
 from opendbc.sunnypilot.car.chrysler.fingerprints_ext import FW_VERSIONS_EXT
 
 Ecu = CarParams.Ecu
