+++ opendbc/car/hyundai/fingerprints.py
@@ -2,7 +2,7 @@
 from opendbc.car.structs import CarParams
 from opendbc.car.hyundai.values import CAR
 
-from opendbc.sunnypilot.car.fw_versions_ext import merge_fw_versions
+from opendbc.sunnypilot.car.fingerprints_ext import merge_fw_versions
 from opendbc.sunnypilot.car.hyundai.fingerprints_ext import FW_VERSIONS_EXT
 
 Ecu = CarParams.Ecu
